Release/v1.0

+++ lib/domain/controllers/discussions/discussion_item_controller.dart
@@ -31,22 +31,27 @@
  */
 
 import 'package:easy_localization/easy_localization.dart';
+import 'package:event_hub/event_hub.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
 import 'package:projects/data/models/from_api/discussion.dart';
 import 'package:projects/data/services/discussion_item_service.dart';
+import 'package:projects/data/services/project_service.dart';
 import 'package:projects/domain/controllers/comments/item_controller/discussion_comment_item_controller.dart';
 import 'package:projects/domain/controllers/comments/new_comment/new_discussion_comment_controller.dart';
 import 'package:projects/domain/controllers/discussions/actions/discussion_editing_controller.dart';
 import 'package:projects/domain/controllers/discussions/discussions_controller.dart';
+import 'package:projects/domain/controllers/messages_handler.dart';
 import 'package:projects/domain/controllers/navigation_controller.dart';
-import 'package:projects/domain/controllers/projects/detailed_project/detailed_project_controller.dart';
 import 'package:projects/domain/controllers/projects/detailed_project/project_discussions_controller.dart';
 import 'package:projects/domain/controllers/user_controller.dart';
 import 'package:projects/internal/locator.dart';
+import 'package:projects/internal/utils/debug_print.dart';
+import 'package:projects/presentation/shared/widgets/styled/styled_alert_dialog.dart';
 import 'package:projects/presentation/views/discussions/creating_and_editing/discussion_editing/discussion_editing_screen.dart';
 import 'package:projects/presentation/views/discussions/creating_and_editing/discussion_editing/select/manage_discussion_subscribers_screen.dart';
 import 'package:projects/presentation/views/discussions/widgets/discussion_status_BS.dart';
+import 'package:projects/presentation/views/project_detailed/project_detailed_view.dart';
 import 'package:projects/presentation/views/task_detailed/comments/new_comment_view.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
 import 'package:visibility_detector/visibility_detector.dart';
@@ -114,7 +119,7 @@ class DiscussionItemController extends GetxController {
   }
 
   Future<void> updateMessageStatus(int newStatus) async {
-    var newStatusStr = newStatus == 1 ? tr('archived') : tr('open');
+    var newStatusStr = newStatus == 1 ? 'archived' : 'open';
 
     try {
       Discussion result = await _api.updateMessageStatus(
@@ -150,25 +155,35 @@ class DiscussionItemController extends GetxController {
     }
   }
 
-  Future<void> deleteMessage() async {
-    try {
-      Discussion result = await _api.deleteMessage(id: discussion.value.id);
-      if (result != null) {
-        Get.back();
-        await Get.find<DiscussionsController>().loadDiscussions();
-        //TODO refactoring needed
+  Future<void> deleteMessage(context) async {
+    await Get.dialog(StyledAlertDialog(
+      titleText: tr('deleteDiscussionTitle'),
+      contentText: tr('deleteDiscussionAlert'),
+      acceptText: tr('delete').toUpperCase(),
+      onCancelTap: () async => Get.back(),
+      onAcceptTap: () async {
         try {
-          // ignore: unawaited_futures
-          Get.find<ProjectDetailsController>().refreshData();
-          // ignore: unawaited_futures
-          Get.find<ProjectDiscussionsController>().loadProjectDiscussions();
+          Discussion result = await _api.deleteMessage(id: discussion.value.id);
+          if (result != null) {
+            Get.back();
+            Get.back();
+            MessagesHandler.showSnackBar(
+                context: context, text: tr('discussionDeleted'));
+            await Get.find<DiscussionsController>().loadDiscussions();
+            //TODO refactoring needed
+            try {
+              locator<EventHub>().fire('needToRefreshProjects');
+              // ignore: unawaited_futures
+              Get.find<ProjectDiscussionsController>().loadProjectDiscussions();
+            } catch (e) {
+              printError(e);
+            }
+          }
         } catch (e) {
-          debugPrint(e);
+          printError(e);
         }
-      }
-    } catch (e) {
-      debugPrint(e);
-    }
+      },
+    ));
   }
 
   void handleVisibilityChanged(VisibilityInfo info) {
@@ -208,5 +223,17 @@ class DiscussionItemController extends GetxController {
         'onConfirm': () => controller.confirm(context),
       },
     );
+  }
+
+  void toProjectOverview() async {
+    var projectService = locator<ProjectService>();
+    var project = await projectService.getProjectById(
+      projectId: discussion.value.projectOwner.id,
+    );
+    if (project != null)
+      Get.find<NavigationController>().to(
+        ProjectDetailedView(),
+        arguments: {'projectDetailed': project},
+      );
   }
 }
